Move timers to separate file
